requirements.txt
flake8
pytest
python-hcl2